+++ src/test/java/com/pingcap/tikv/GrpcUtils.java
@@ -17,10 +17,10 @@ package com.pingcap.tikv;
 
 import com.google.common.collect.Lists;
 import com.google.protobuf.ByteString;
+import com.pingcap.tikv.codec.Codec.BytesCodec;
 import com.pingcap.tikv.codec.CodecDataOutput;
 import com.pingcap.tikv.kvproto.Metapb.*;
 import com.pingcap.tikv.kvproto.Pdpb.*;
-import com.pingcap.tikv.types.BytesType;
 import java.util.Arrays;
 
 public class GrpcUtils {
@@ -55,7 +55,7 @@ public class GrpcUtils {
 
   public static ByteString encodeKey(byte[] key) {
     CodecDataOutput cdo = new CodecDataOutput();
-    BytesType.writeBytes(cdo, key);
+    BytesCodec.writeBytes(cdo, key);
     return cdo.toByteString();
   }
